refactor: extract theme helpers to :common
Prep for moving 'Themes' to :common
-> Prep for moving 'AnkiActivity' to :common

Part of 20737
