+++ AnkiDroid/src/main/java/com/ichi2/anki/AnkiActivity.kt
@@ -66,6 +66,7 @@ import com.ichi2.anki.android.input.shortcut
 import com.ichi2.anki.common.annotations.LegacyNotifications
 import com.ichi2.anki.common.annotations.NeedsTest
 import com.ichi2.anki.common.crashreporting.CrashReportService
+import com.ichi2.anki.common.utils.android.getColorFromAttr
 import com.ichi2.anki.common.utils.android.showThemedToast
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.compat.CompatHelper
@@ -668,7 +669,7 @@ open class AnkiActivity(
     fun setNavigationBarColor(
         @AttrRes attr: Int,
     ) {
-        window.navigationBarColor = Themes.getColorFromAttr(this, attr)
+        window.navigationBarColor = getColorFromAttr(this, attr)
     }
 
     fun closeCollectionAndFinish() {
