+++ common/android/build.gradle.kts
@@ -29,6 +29,7 @@ dependencies {
 
     implementation(libs.androidx.annotation)
     implementation(libs.androidx.core.ktx)
+    implementation(libs.google.material)
     implementation(libs.jakewharton.timber)
     implementation(libs.kotlinx.coroutines.core)
